+++ CHANGES.md
@@ -4,6 +4,12 @@
 - Improved functions execution times
 - Improved functions execution times
 - Create a new builds worker to handle building of deployments
+- **[ Breaking ]** **Tags** have been renamed to **Deployments**
+  - Rename `tagId` to `deplyomentId` in collections
+  - Rename tags to deployments in the docs
+  - Updated endpoints to reflect the new terminology 
+  - Updated UI with these changes
+  - Updated event names from `function.tags.*` to `function.deployments.*`
 # Version 0.12.1
 
 ## Bugs